+++ docs/doxygen/html/search/all_d.js
@@ -1,14 +1,14 @@
 var searchData=
 [
-  ['param',['param',['../classhal_1_1maths_1_1plotter_1_1_plot2d.html#a60f7295d6cad1639e30c6b0407a2b983',1,'hal.maths.plotter.Plot2d.param()'],['../classhal_1_1maths_1_1plotter_1_1_plot3d.html#ad17823fc8ceb91fa2c602c45f8eae70e',1,'hal.maths.plotter.Plot3d.param()'],['../classhal_1_1maths_1_1plotter_1_1_plot4d.html#a90b39427c236100f8c080e08e2222f2d',1,'hal.maths.plotter.Plot4d.param()']]],
+  ['param',['param',['../classhal_1_1charts_1_1plotter_1_1_plot2d.html#a66802c3c0beadbe0ccef615620ec1e36',1,'hal.charts.plotter.Plot2d.param()'],['../classhal_1_1charts_1_1plotter_1_1_plot3d.html#acbc65e1bdff5440a853c4d5c331907af',1,'hal.charts.plotter.Plot3d.param()'],['../classhal_1_1charts_1_1plotter_1_1_plot4d.html#a925008c5f9e89231a632fdad42a6cdff',1,'hal.charts.plotter.Plot4d.param()']]],
   ['parse',['parse',['../classhal_1_1internet_1_1parser_1_1_html_table.html#addcc7238344dce5f6f71426ae03ce362',1,'hal::internet::parser::HtmlTable']]],
   ['parse_5fdata',['parse_data',['../classhal_1_1ml_1_1data_1_1parser_1_1_c_s_v_parser.html#ac47a4070d172a1ada61060610749f5e8',1,'hal::ml::data::parser::CSVParser']]],
   ['parse_5fquery',['parse_query',['../classhal_1_1internet_1_1engines_1_1_search_engine.html#afde9fa1646b6057ecd94746132c9b7c1',1,'hal::internet::engines::SearchEngine']]],
   ['parse_5furl',['parse_url',['../classhal_1_1internet_1_1web_1_1_webpage.html#af6e3dcd4dbebe8e1b196fded24e4025e',1,'hal::internet::web::Webpage']]],
   ['parser',['Parser',['../classhal_1_1ml_1_1data_1_1parser_1_1_parser.html',1,'hal::ml::data::parser']]],
-  ['plot',['plot',['../classhal_1_1maths_1_1plotter_1_1_plot2d.html#aa340a489f84fc82aa7c69a868ebd11c3',1,'hal.maths.plotter.Plot2d.plot()'],['../classhal_1_1maths_1_1plotter_1_1_plot3d.html#a1e249de9ab5dac1acc94f82bada6deb0',1,'hal.maths.plotter.Plot3d.plot()'],['../classhal_1_1maths_1_1plotter_1_1_plot4d.html#aa776ce05b0f159953ea817402aa9e280',1,'hal.maths.plotter.Plot4d.plot()']]],
-  ['plot2d',['Plot2d',['../classhal_1_1maths_1_1plotter_1_1_plot2d.html',1,'hal::maths::plotter']]],
-  ['plot3d',['Plot3d',['../classhal_1_1maths_1_1plotter_1_1_plot3d.html',1,'hal::maths::plotter']]],
-  ['plot4d',['Plot4d',['../classhal_1_1maths_1_1plotter_1_1_plot4d.html',1,'hal::maths::plotter']]],
+  ['plot',['plot',['../classhal_1_1charts_1_1plotter_1_1_plot2d.html#a0fa5efc521db379607a78abc972e34d6',1,'hal.charts.plotter.Plot2d.plot()'],['../classhal_1_1charts_1_1plotter_1_1_plot3d.html#a56f281f0b9fb9a0f37247c5a0f4791df',1,'hal.charts.plotter.Plot3d.plot()'],['../classhal_1_1charts_1_1plotter_1_1_plot4d.html#a0826f5d596a40fb6ec31f6f2636b55e9',1,'hal.charts.plotter.Plot4d.plot()']]],
+  ['plot2d',['Plot2d',['../classhal_1_1charts_1_1plotter_1_1_plot2d.html',1,'hal::charts::plotter']]],
+  ['plot3d',['Plot3d',['../classhal_1_1charts_1_1plotter_1_1_plot3d.html',1,'hal::charts::plotter']]],
+  ['plot4d',['Plot4d',['../classhal_1_1charts_1_1plotter_1_1_plot4d.html',1,'hal::charts::plotter']]],
   ['prettify',['prettify',['../classhal_1_1files_1_1models_1_1_file_system.html#a81206c773e9176fb2c0537a731381b7a',1,'hal::files::models::FileSystem']]]
 ];