+++ docs/doxygen/html/search/all_f.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['scatter',['scatter',['../classhal_1_1maths_1_1plotter_1_1_plot2d.html#adff2970c2337d309b1943479d216d2f6',1,'hal.maths.plotter.Plot2d.scatter()'],['../classhal_1_1maths_1_1plotter_1_1_plot3d.html#ac426f777196d1c85ae841e59ea43ad89',1,'hal.maths.plotter.Plot3d.scatter()'],['../classhal_1_1maths_1_1plotter_1_1_plot4d.html#a1f822a4cfb97006651da35cb2f5e55d1',1,'hal.maths.plotter.Plot4d.scatter()']]],
+  ['scatter',['scatter',['../classhal_1_1charts_1_1plotter_1_1_plot2d.html#af645ee6b4deb16c8a773b7f40909a1d2',1,'hal.charts.plotter.Plot2d.scatter()'],['../classhal_1_1charts_1_1plotter_1_1_plot3d.html#a70697dd4170ed79cf48d0835b8e38a29',1,'hal.charts.plotter.Plot3d.scatter()'],['../classhal_1_1charts_1_1plotter_1_1_plot4d.html#aa5b31430953600e79cdb0dd609d0e971',1,'hal.charts.plotter.Plot4d.scatter()']]],
   ['searchengine',['SearchEngine',['../classhal_1_1internet_1_1engines_1_1_search_engine.html',1,'hal::internet::engines']]],
   ['searchengineresult',['SearchEngineResult',['../classhal_1_1internet_1_1engines_1_1_search_engine_result.html',1,'hal::internet::engines']]],
   ['seleniumform',['SeleniumForm',['../classhal_1_1internet_1_1selenium_1_1_selenium_form.html',1,'hal::internet::selenium']]],